# api-dh-v2/static/swagger.json
﻿{
  "openapi": "3.0.0",
  "info": {
    "description": "GPA Meu desconto API",
    "version": "1.0.0",
    "title": "GPA Meu desconto API",
    "contact": {
      "email": "@dunnhumby.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "host": "",
  "servers": [
      {
        "url": "/"
      }
    ],
...
    "paths": {
      "/getMultipleAudience": {
        "post": {
          "tags": [
              "Obtém audiências sem autenticação"
          ],
          "description": "De acordo com os Parâmetros utilizados retorna a audiência ou audiências <br><br> *Parâmetros:* getMultipleAudienceRequestBody<br> Os parâmetros não são todos obrigatórios e afetam o comportamento da API. Os parâmetros estão descritos nos schemas:<br> *Request:* getMultipleAudienceRequestBody <br> *Retorno :* getMultipleAudienceResponse <br>*Retorno de Erro :* getMultipleAudienceErrorResponse",
          "summary": "Obtém audiência ou audiências calculadas de acordo com os parâmetros utilizados",
          "requestBody": {
              "description": "getMultipleAudience Request Object",
              "required": true,
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/getMultipleAudienceRequestBody"
                  }
                }
              }
            },
            "produces": [
              "application/json"
            ],
          "responses": {
            "200": {
              "description": "Success",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/getMultipleAudienceResponse"
                  }
                }
              }
            },
            "500": {
              "description": "Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/getMultipleAudienceErrorResponse"
                  }
                }
              }
            }
          }
        },
        "get": {
          "tags": [
            "Obtém audiências sem autenticação"
          ],
          "description": "De acordo com os parâmetros utilizados retorna a audiência ou audiências <br><br> *Parâmetros:* getMultipleAudienceRequestBody<br> Os parâmetros não são todos obrigatórios e afetam o comportamento da API. Os parâmetros estão descritos nos schemas:<br> *Request:* getMultipleAudienceRequestBody <br> *Retorno :* getMultipleAudienceResponse <br>*Retorno de Erro :* getMultipleAudienceErrorResponse",
          "summary": "Obtém audiência ou audiências calculadas de acordo com os parâmetros utilizados",

          "requestBody": {
              "description": "getMultipleAudience Request Object",
              "required": true,
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/getMultipleAudienceRequestBody"
                  }
                }
              }
            },
            "produces": [
              "application/json"
            ],
          "responses": {
            "200": {
              "description": "Success",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/getMultipleAudienceResponse"
                  }
                }
              }
            },
            "500": {
              "description": "Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/getMultipleAudienceErrorResponse"
                  }
                }
              }
            }
          }
        },
        "options": {
          "tags": [
              "Obtém audiências sem autenticação"
          ],
          "description": "De acordo com os parâmetros utilizados retorna a audiência ou audiências <br><br> *Parâmetros:* getMultipleAudienceRequestBody<br> Os parâmetros não são todos obrigatórios e afetam o comportamento da API. Os parâmetros estão descritos nos schemas:<br> *Request:* getMultipleAudienceRequestBody <br> *Retorno :* getMultipleAudienceResponse <br>*Retorno de Erro :* getMultipleAudienceErrorResponse",
          "summary": "Obtém audiência ou audiências calculadas de acordo com os parâmetros utilizados",

          "requestBody": {
              "description": "getMultipleAudience Request  Object",
              "required": true,
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/getMultipleAudienceRequestBody"
                  }
                }
              }
            },
            "produces": [
              "application/json"
            ],
          "responses": {
            "200": {
              "description": "Success",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/getMultipleAudienceResponse"
                  }
                }
              }
            },
            "500": {
              "description": "Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/getMultipleAudienceErrorResponse"
                  }
                }
              }
            }
          }
        }

      },
      "/getMultipleAudienceAuth": {
        "post": {
          "security": [
            {
              "api_key": []
            }
          ],
            "tags": [
              "Obtém audiências com autenticação"
          ],
          "description": "De acordo com os parâmetros utilizados retorna a audiência ou audiências <br><br> *Parâmetros:* getMultipleAudienceRequestBody<br> Os parâmetros não são todos obrigatórios e afetam o comportamento da API. Os parâmetros estão descritos nos schemas:<br> *Request:* getMultipleAudienceRequestBody <br> *Retorno :* getMultipleAudienceResponse <br>*Retorno de Erro :* getMultipleAudienceErrorResponse",
          "summary": "Obtém audiência ou audiências calculadas de acordo com os parâmetros utilizados",


          "requestBody": {
              "description": "getMultipleAudienceAuth Request Post Object",
              "required": true,
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/getMultipleAudienceRequestBody"
                  }
                }
              }
            },
            "produces": [
              "application/json"
            ],

          "responses": {
            "200": {
              "description": "Success",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/getMultipleAudienceResponse"
                  }
                }
              }
            },
            "500": {
              "description": "Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/getMultipleAudienceErrorResponse"
                  }
                }
              }
            }
          }
        }
      },
      "/getReportDataAuth": {
        "post": {
          "security": [
            {
              "api_key": []
            }
          ],
            "tags": [
              "Obtém dados para o report"
          ],
          "description": "De acordo com os parâmetros utilizados retorna os dados de campanha de acordo com os parâmetros *Parâmetros:* getReportDataAuthBody <br>   *Retorno :* getReportDataAuthResponse <br>*Retorno de Erro :* getReportDataAuthErrorResponse",
          "summary": "Obtém dados de campanha de acordo com o fornecedor",
          "requestBody": {
              "description": "getReportDataAuth Request Post Object",
              "required": true,
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/getReportDataAuthBody"
                  }
                }
              }
            },
            "produces": [
              "application/json"
            ],

          "responses": {
            "200": {
              "description": "Success",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/getReportDataAuthResponse"
                  }
                }
              }
            },
            "500": {
              "description": "Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/getReportDataAuthErrorResponse"
                  }
                }
              }
            }
          }
        }
      }
...
      "/getIndicadorAuth": {
        "post": {
          "security": [
            {
              "api_key": []
            }
          ],
            "tags": [
              "Obtém dados de indicadores Meu Desconto"
          ],
          "description": "Traz dados de indicadores para exibição na pagina inicial do Meu Desconto *Parâmetros:* getReportDataAuthBody <br>   *Retorno :* getIndicadorAuthResponse <br>*Retorno de Erro :* getIndicadorAuthErrorResponse",
          "summary": "Obtém dados de indicadores do meu desconto, chamada uma vez por semana",
          "produces": [
              "application/json"
            ],

          "responses": {
            "200": {
              "description": "Success",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/getIndicadorAuthResponse"
                  }
                }
              }
            },
            "500": {
              "description": "Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/getIndicadorAuthErrorResponse"
                  }
                }
              }
            }
          }
        }
      },
...
    },
    "components": {
      "securitySchemes":{
        "api_key":  {

        "type": "http",
        "scheme": "basic",
        "name" : "authorization"
        }
       },
      "schemas": {
          "getMultipleAudienceRequestBody": {
          "type": "object",
          "properties": {
            "audience":
            {
              "type": "array",
              "items":
                { "$ref": "#/components/schemas/getAudienceBody"}
            }

          }
        },
        "getAudienceBody": {
          "type": "object",
          "properties": {
            "id": {
              "required": "false",
              "type": "number",
              "format": "number"
            },
            "modalidade": {
              
              "type":"array",
              "required": "false",
              "description": "*Filtros necessários para busca de audiência. Pode ser por Categoria,grupo,sub-grupo,familia ou produto.<br> no caso de produto poderá ser enviado uma lista de produtos ", 
              "$ref": "#/components/schemas/getModalidade"

            
          }, 
          "duracao": {
              "type": "integer",
              "required": "false",
              "description": "*Codigo de ciclo de alocação.*<br> 1 - Primeira semana  dias <br> 2 - Segunda Semana <br> 3 - 14 dias "
          },

          "objetivo": {
              "type": "string",
              "required": "false",
              "description": "*Objetivo da campanha.*<br> Utilizado para calcular a taxa de conversão(coeficiente) caso não exista no arquivo redrates.csv. Também é utilizado filtro<br> valores admitidos:<br> 'frequency': ' is_frequency_increase_candidate  <br>'loyal_a': '  is_loyal_a  <br>'loyal_b': '  is_loyal_b  <br>'ticket': '  is_ticket_increase_candidate  <br> 'penetration':  is_product_increase_candidate_a  <br>'recover':  is_recovery_candidate  <br>'launch':  is_new_product_candidate  '"
          } ,

          "pFidelidade": {
              "type": "string",
              "format": "string",
              "required": "false",
              "description": "*Tipo de fidelidade*: Valores admitidos : PA+,PA, EX. O comportamento padrao caso não seja enviado é EX",
              "enum:": [
                  "PA",
                  "EX", "PA+"
              ]
          },
  
          "unidadesMaxProduto": {
              "type": "integer",
              "format": "integer",
              "required": "false",
              "description": "*Quantidade de Unidades máximas do produto.*<br> Limita a quantidade de unidades por produto. Os cálculos utilizarão o menor número entre o Parâmetro e configuração de unidades por visita do L20"

          },
    
            "mecanica": {
              "type":"object",
              "required": "false",
              "description": "*Dicionário com os tipo de mecânica a ser utilizado*<br> definido no schema getMecanica<br> caso não seja passada a mecânica(ou tipo inexistente) o investimento máximo e faturamento serão considerados ZERO ",
              "items":{
              "$ref": "#/components/schemas/getMecanica"
              }
            },
            "regioes": {
              "type": "array",
              "required": "false",
              "description": "*Matriz com as regiões a serem consideradas*. <br> Enviar lista com as siglas dos estados.",
              "items": {
                  "type": "string"
              }

          },

          "filtros": {

                "$ref": "#/components/schemas/getMultipleAudienceRequestFilters"

            }


          }
        },
...
        "getMecanica": {
          "type": "object",
          "required": "false",
          "description": "*Dicionário com a mecânica*. ",
          "properties": {
              "tipoMecanica": {
                  "required": "true",
                  "description": "*Tipo de mecânica a ser utilizado*. <br> Os seguintes valores são admitidos:<br> 1. LEVE_X_PAGUE_Y <br> 2. PERCENTUAL_X_NA_Y_UNIDADE<br> 3. PERCENTUAL_X<br> 4. PERCENTUAL<br> 5. ABSOLUTO_X<br> 6. ABSOLUTO ",
                  "type": "string"
              },
              "valorDesconto": {
                  "required": "false",
                  "description": "*Valor do desconto para a mecânica.* utilizado nos tipos de mecânica:<br> 3. PERCENTUAL_X<br> 4. PERCENTUAL<br> 5. ABSOLUTO_X<br> 6. ABSOLUTO ",
                  "type": "number"
              },
              "unidadesPagas": {
                  "required": "false",
                  "description": "*número de unidades pagas em mecânicas do tipo LEVE X PAGUE Y.* utilizado nos tipos de mecânica:<br> 1. LEVE_X_PAGUE_Y<br> ",
                  "type": "integer"
              },
              "unidadesLevadas": {
                  "required": "false",
                  "description": "*número de unidades levadas em mecânicas do tipo LEVE X PAGUE Y.* utilizado nos tipos de mecânica:<br> 1. LEVE_X_PAGUE_Y<br> ",
                  "type": "integer"
              },
              "descontoEmOutraUnidade": {
                  "required": "false",
                  "description": "*Desconto em unidades por um certo número de unidades compradas.* utilizado nos tipos de mecanica:<br> 2. PERCENTUAL_X_NA_Y_UNIDADE<br> ",
                  "type": "number"
              },
              "unidade": {
                  "required": "false",
                  "description": "*número de unidades compradas a partir do qual o desconto é aplicado.* utilizado nos tipos de mecânica:<br> 2. PERCENTUAL_X_NA_Y_UNIDADE<br> ",
                  "type": "number"
              }
          }
      } ,
...
        "getMultipleAudienceErrorResponse": {
          "type": "object",
          "description": "Ao ocorrer um erro será enviada uma resposta padrão com os seguintes valores : <br>estFaturamento: 0<br> estPublicoAlocado: 0<br> estInvestimentoMaximo: 0<br> valoraudiência: 0<br>qtdMediaCliente: 1.0<br> MinDiscountPercentual: 0<br> MinDiscountAbs: 0<br> MaxDiscountAbs: 0<br>pesoVariavel: false<br> MaxNumberUN: 3<br> MinNumberUN: 1<br>allowUnder18: true ",
          "properties": {
            "estFaturamento": {
              "type": "number",
              "format": "number"
            },
            "estPublicoAlocado": {
              "type": "number",
              "format": "number"
            },
            "estInvestimentoMaximo": {
              "type": "number",
              "format": "number"
            },
            "valoraudiência": {
              "type": "number",
              "format": "number"
            },
            "qtdMediaCliente": {
              "type": "number",
              "format": "number"
            },
            "MinDiscountPercentual": {
              "type": "number",
              "format": "number"
            },
            "MinDiscountAbs": {
              "type": "number",
              "format": "number"
            },
            "MaxDiscountAbs": {
              "type": "number",
              "format": "number"
            },
            "pesoVariavel": {
              "type": "boolean",
              "format": "boolean"
            },
            "MaxNumberUN": {
              "type": "number",
              "format": "number"
            },
            "MinNumberUN": {
              "type": "number",
              "format": "number"
            },
            "allowUnder18": {
              "type": "boolean",
              "format": "boolean"
            },
            "pFidelidade": {
              "type": "boolean",
              "format": "boolean"
            },

            "id": {
              "type": "number",
              "format": "number"
            },
            "Razao": {
              "type": "string",
              "format": "string",
              "description": "*Caso tenha ocorrido algum erro, envia uma mensagem com a razao do ocorrido*: <br> "
            }
          }
        },
...
        "getMultipleAudienceResponse": {
          "type": "object",
          "description": "*Resposta do método getMultipleAudience*. ",
          "properties": {
            "estFaturamento": {
              "type": "number",
              "format": "number",
              "description": "*Valor calculado*: Fórmula: (preco do produto - Desconto por unidade) * estimativa publico alocado * quantidade média de unidades<br> para os tipos de mecanica PERCENTUAL_X_NA_Y_UNIDADE,LEVE_X_PAGUE_Y o formula é: <br>(preco do produto - Desconto por unidade) * estimativa publico alocado * (quantidade média de unidades/número de unidades)"
          },
            "estPublicoAlocado": {
              "type": "number",
              "format": "number",
              "description": "*Valor calculado*: valoraudiência * coeficiente<br> o coeficiente é taxa de conversão do csv redrates.utilizand os filtros pfidelidade, aba, level_code, L20_code ou L21_code. caso não encontre utiliza a tabela abaixo dependendo do objetivo.<br>'frequency': se pFidelidade = 'PA'  0.0116 senão 0.0073<br>'loyal_a': se pFidelidade = 'PA' 0.0116 senão 0.0073<br>'loyal_b': se pFidelidade = 'PA' 0.0116 senão 0.0073 'ticket': se pFidelidade = 'PA' 0.0116 senão 0.0073<br>'penetration': 0.0010<br> 'recover': 0.0010<br> 'launch': 0.0010 <br> <h4> caso o objetivo não tenha sido passado o coeficiente será se pFidelidade = 'PA' 0.0116 senão 0.0073 </h4>"
          },
            "estInvestimentoMaximo": {
              "type": "number",
              "format": "number",
              "description": "*Valor Calculado:*  Formula: Desconto por unidade x estimativa publico alocado x quantidade média de unidades<br> para os tipos de mecanica PERCENTUAL_X_NA_Y_UNIDADE,LEVE_X_PAGUE_Y o formula é: <br> Desconto por unidade * estimativa publico alocado * (quantidade média de unidades/número de unidades)"
          },
            "valoraudiência": {
              "type": "number",
              "format": "number",
              "description": "*Retorna contagem do cubo de audiência de acordo com os Parâmetros passados para a API*"
          },
            "qtdMediaCliente": {
              "type": "number",
              "format": "number",
              "description": "*Quantidade média do produto comprado por visita.:* Retorna o valor cadastrado para o produto ou L20 na tabela index_l21. "
          },
            "MinDiscountPercentual": {
              "type": "number",
              "format": "number",
              "description": "*Percentual do valor minimo de desconto.:* Retorna o valor cadastrado para o produto ou L20 na tabela index_l21. "
          },
            "MinDiscountAbs": {
              "type": "number",
              "format": "number",
              "description": "*Valor mínimo do desconto. Calculado a partir o do preço e percentual mínimo de desconto*. "
          },
            "MaxDiscountAbs": {
              "type": "number",
              "format": "number",
              "description": "*Valor máximo do desconto. 50% do preço do produto*. "
          },
            "pesoVariavel": {
              "type": "boolean",
              "format": "boolean",
              "description": "*Flag que indica se o produto tem peso variavel. Retorna true ou false*. "
          },
            "MaxNumberUN": {
              "type": "number",
              "format": "number",
              "description": "*Valor máximo de unidades de necessidade. Configurado para retornar sempre 3*. "
          },
            "MinNumberUN": {
              "type": "number",
              "format": "number",
              "description": "*Valor mínimo de unidades de necessidade. Configurado para retornar sempre 1*. "
          },
            "allowUnder18": {
              "type": "boolean",
              "format": "boolean",
              "description": "*Retorna se o produto é permitido para menores de 18 anos. Se o produto ou l20 for alcool retorna false, caso contrário true* "

            },
            "pFidelidade": {
              "type": "boolean",
              "format": "boolean",
              "description": "*Tipo de fidelidade passado como parâmetro*: Valores admitidos : PA, EX, PA+",
              "enum:":["PA","EX","PA+"]
          },

            "id": {
              "type": "number",
              "format": "number",
              "description": "*Retorna o campo id caso tenha sido passado como parâmetro* "
          },
          "Razao": {
              "type": "string",
              "format": "string",
              "description": "*Caso tenha ocorrido algum erro, envia uma mensagem com a razao do ocorrido* <br> se o produto e l20 não estiver carregado nos cubos envia a seguinte mensagem:<br> Produto ({0}) e l20 ({1}) nao encontrados! "
          }

          }
        },
        "getReportDataAuthBody": {
          "type": "object",
          "description": "",
          "properties": {
            "vendor_class_code": {
              "required": "false",
              "description": "*. Codigo do fabricante  ex:5498  'Schincariol', para GPA Deixar em branco",
              "type": "string"
                
              },
          "vendor_code": {
              "required": "false",
              "description": "*Código do Distribuidor ex :00000000028067  'HNK BR BEBIDAS LTDA.'. para GPA deixar em branco",
              "type": "string"
                
              },
          "regioes": {
                "type": "array",
                "required": "false",
                "description": "*Matriz com as regiões a serem consideradas*. <br> Enviar lista com as siglas dos estados.",
                "items": {
                    "type": "string"
                }  
          },
          "bandeira": {
            "type": "string",
            "format": "string",
            "required": "false",
            "description": "*Tipo de fidelidade*: Valores admitidos :PA, EX. O comportamento padrao caso não seja enviado é cálculo para as duas bandeiras juntas",
            "enum:": [
                "PA",
                "EX"
            ]
          }
        }
      },
        "getIndicadorAuthResponse": {
          "type": "object",
          "description": "",
          "properties": {
            "Plataforma": {
              "type": "array",
              "required": "false",
              "items": {
                "$ref": "#/components/schemas/getIndicadorPlataformaDataResponse"
              }  
            },
            "Industria": {
              "type": "array",
              "required": "false",
              "items": {
                "$ref": "#/components/schemas/getIndicadorIndustriaDataResponse"
              }  
            }
          }
        }
...
        "getIndicadorPlataformaDataResponse": {
          "type": "object",
          "description": "",
          "properties": {
            "bandeira": {
              "type": "string"
            },
            "ofertaAtivadas": {
              "type": "number"
            },
            "clientesComprando": {
              "type": "number"
            },
            "vendas": {
              "type": "number"
            }
          
          }
        },
        "getIndicadorIndustriaDataResponse": {
          "type": "object",
          "description": "",
          "properties": {
            "bandeira": {
              "type": "string"
            },
            "vendor_code": {
              "type": "string"
            },
            "vendor_class_code": {
              "type": "string"
            },
            "ofertaAtivadas": {
              "type": "number"
            },
            "clientesComprando": {
              "type": "number"
            },
            "clientesUnicos": {
              "type": "number"
            },
            "vendas": {
              "type": "number"
            }
          
          }
        },
        "getReportDataAuthResponse": {
          "type": "object",
          "description": "",
          "properties": {
            "fornecedor": {
              "type": "array",
              "required": "false",              
              "items": {
                "$ref": "#/components/schemas/getReportFornecedorDataResponse"
              }  
            }
          }
        },
        "getReportFornecedorDataResponse": {
          "type": "object",
          "description": "",
          "properties": {
            "nome": {
              "type": "string"
            },
            "dados":
            {
              "$ref": "#/components/schemas/getReportDataResponse"
            }
          }
        },
...
      "getReportDataAuthErrorResponse": {
          "type": "object",
          "description": "",
          "properties": {
            "Razao": {
              "type": "string",
              "description":"* motivo do erro *"
            }
          }
        },
...
      }
    }
}